Se cambia tipo de dato y se corrige BaseDao

+++ src/main/java/org/jugni/apps/pico/DAO/CuentaTipoDao.java
@@ -3,7 +3,7 @@ package org.jugni.apps.pico.DAO;
 import org.hibernate.SessionFactory;
 import org.jugni.apps.pico.modelos.CuentaTipo;
 
-public class CuentaTipoDao extends BaseDao<CuentaTipo, Byte> {
+public class CuentaTipoDao extends BaseDao<CuentaTipo, Short> {
 
   public CuentaTipoDao(SessionFactory factory) {
     super(factory, CuentaTipo.class);
